+++ packages/smarthr-ui/src/components/ComboBox/MultiComboBox/MultiComboBox.tsx
@@ -23,7 +23,7 @@ import { FaCaretDownIcon } from '../../Icon'
 import { areComboBoxItemsEqual } from '../comboBoxHelper'
 import { useFocusControl } from '../useFocusControl'
 import { useListBox } from '../useListBox'
-import { useOptions } from '../useOptions'
+import { useMultiOptions } from '../useOptions'
 
 import { MultiSelectedItem } from './MultiSelectedItem'
 import { hasParentElementByClassName } from './multiComboBoxHelper'
@@ -176,7 +176,7 @@ const ActualMultiComboBox = <T,>(
   const [uncontrolledInputValue, setUncontrolledInputValue] = useState('')
   const inputValue = isInputControlled ? controlledInputValue : uncontrolledInputValue
   const [isComposing, setIsComposing] = useState(false)
-  const { options } = useOptions({
+  const { options } = useMultiOptions({
     items,
     selected: selectedItems,
     creatable,
@@ -449,6 +449,7 @@ const ActualMultiComboBox = <T,>(
   ])
 
   return (
+    // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
     <div
       {...wrapperStyleProps}
       ref={outerRef}