+++ packages/smarthr-ui/src/components/Dialog/FormDialog/FormDialogContentInner.tsx
@@ -9,6 +9,7 @@ import React, {
 import { tv } from 'tailwind-variants'
 
 import { type DecoratorsType } from '../../../hooks/useDecorators'
+import { type ResponseMessageType, useResponseMessage } from '../../../hooks/useResponseMessage'
 import { Button } from '../../Button'
 import { Cluster, Stack } from '../../Layout'
 import { ResponseMessage } from '../../ResponseMessage'
@@ -16,8 +17,6 @@ import { Section } from '../../SectioningContent'
 import { DialogBody, Props as DialogBodyProps } from '../DialogBody'
 import { DialogHeader, type Props as DialogHeaderProps } from '../DialogHeader'
 import { dialogContentInner } from '../dialogInnerStyle'
-
-import type { ResponseMessageType } from '../../../types'
 
 export type BaseProps = PropsWithChildren<
   DialogHeaderProps &
@@ -86,30 +85,7 @@ export const FormDialogContentInner: FC<FormDialogContentInnerProps> = ({
     [onSubmit, onClickClose],
   )
 
-  const calculatedResponseStatus = useMemo(() => {
-    if (!responseMessage) {
-      return {
-        isProcessing: false,
-        visibleMessage: false,
-      }
-    }
-
-    if (responseMessage.status === 'processing') {
-      return {
-        isProcessing: true,
-        visibleMessage: false,
-      }
-    }
-
-    return {
-      isProcessing: false,
-      visibleMessage: true,
-      // HINT: statusがprocessingではない === success or errorであることが確定する
-      // success or error の場合、text属性も必ず存在する
-      status: responseMessage.status as 'success' | 'error',
-      message: (responseMessage as { text: string }).text,
-    }
-  }, [responseMessage])
+  const calculatedResponseStatus = useResponseMessage(responseMessage)
 
   const styles = useMemo(() => {
     const { form, wrapper, actionArea, buttonArea, message } = formDialogContentInner()
@@ -124,7 +100,7 @@ export const FormDialogContentInner: FC<FormDialogContentInnerProps> = ({
   }, [])
 
   return (
-    // eslint-disable-next-line smarthr/a11y-prohibit-sectioning-content-in-form, smarthr/a11y-heading-in-sectioning-content
+    // eslint-disable-next-line smarthr/a11y-heading-in-sectioning-content, smarthr/a11y-prohibit-sectioning-content-in-form
     <Section className={styles.wrapper}>
       <DialogHeader title={title} subtitle={subtitle} titleTag={titleTag} titleId={titleId} />
       <form onSubmit={handleSubmitAction} className={styles.form}>
@@ -145,7 +121,7 @@ export const FormDialogContentInner: FC<FormDialogContentInnerProps> = ({
               className={styles.buttonArea}
             />
           </Cluster>
-          {calculatedResponseStatus.visibleMessage && (
+          {calculatedResponseStatus.message && (
             <div className={styles.message}>
               <ResponseMessage type={calculatedResponseStatus.status} role="alert">
                 {calculatedResponseStatus.message}